projects/tests/out/test-ml-platform/app/pyproject.toml
[project]
name = "ml-platform"
version = "0.1.0"
description = "FastAPI application generated by NIKAME"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "httpx>=0.26.0,<1.0.0",
    "python-logging-loki>=0.3.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
db = [
    "sqlalchemy>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "Faker>=23.0.0,<24.0.0",
]
cache = ["redis>=5.0.0,<6.0.0"]
messaging = ["aiokafka>=0.10.0,<1.0.0"]
features = ["unleash-client-python>=5.10.0"]
ratelimit = ["fastapi-limiter>=0.1.5,<0.2.0"]
storage = ["boto3>=1.34.0"]
vector = ["qdrant-client>=1.7.0"]
ml = ["mlflow>=2.10.0", "litellm>=1.10.0"]
obs_ml = []


[tool.uv]
managed = true
